feat(file-watching): add file watching to extension
This includes documentation for the custom components as well as
allowing us to watch for changes when the custom component
changes in terms of edit and creation. Using a TS Program
and a watchCompilerHost (while bypassing the compile), we can get
an incremental change handler for ts files in the project using
the tsconfig.json (if available).